+++ libearth/parser/rss2.py
@@ -4,8 +4,6 @@
 Parsing RSS 2.0 feed.
 
 """
-import datetime
-import email.utils
 import re
 
 try:
@@ -13,38 +11,20 @@ try:
 except ImportError:
     from urllib import parse as urlparse
 
-from ..codecs import Rfc3339, Rfc822
-from ..compat import IRON_PYTHON
 from ..compat.etree import fromstring
-from ..feed import (Category, Content, Entry, Feed, Generator, Link,
-                    Person, Text)
-from ..schema import DecodeError
-from ..tz import FixedOffset, guess_tzinfo_by_locale, now, utc
+from ..feed import Category, Entry, Feed, Generator, Link
 from .atom import ATOM_XMLNS_SET
-from .base import ParserBase, SessionBase
+from .base import ParserBase
+from .rss_base import (CONTENT_XMLNS, RSSSession, content_parser,
+                       datetime_parser, guess_default_tzinfo, link_parser,
+                       make_legal_as_atom, person_parser, subtitle_parser,
+                       text_parser)
 from .util import normalize_xml_encoding
 
 
 GUID_PATTERN = re.compile('^(\{{0,1}([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9'
                           'a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12}\}'
                           '{0,1})$')
-CONTENT_XMLNS = 'http://purl.org/rss/1.0/modules/content/'
-
-
-class RSS2Session(SessionBase):
-    """The session class used for parsing the RSS2.0 feed."""
-
-    #: (:class:`str`) The url of the feed to make :class: `~libearth.feed.Link`
-    #: object of which relation is self in the feed.
-    feed_url = None
-
-    #: (:class:`str`) The default time zone name to set the tzinfo of parsed
-    #: :class: `datetime.datetime` object.
-    default_tz_info = None
-
-    def __init__(self, feed_url, default_tz_info):
-        self.feed_url = feed_url
-        self.default_tz_info = default_tz_info
 
 
 rss2_parser = ParserBase()
@@ -60,67 +40,17 @@ def parse_item(element, session):
     return Entry(), session
 
 
-_rfc3339 = Rfc3339()
-_rfc822 = Rfc822()
-_datetime_formats = [
-    ('%Y-%m-%d %H:%M:%S', None),  # daumwebtoon
-    ('%m/%d/%Y %H:%M:%S GMT', utc),  # msdn
-    ('%m/%d/%y %H:%M:%S GMT', utc),  # msdn
-    ('%a, %d %b %Y %H:%M:%S GMT 00:00:00 GMT', utc),  # msdn
-    ('%Y.%m.%d %H:%M:%S', None),  # imbcnews
-    ('%d %b %Y %H:%M:%S %z', None),  # lee-seungjae
-]
-
-
 @parse_channel.path('pubDate', attr_name='updated_at')
 @parse_item.path('pubDate', attr_name='published_at')
 def parse_datetime(element, session):
-    # https://github.com/earthreader/libearth/issues/30
-    string = element.text
-    try:
-        return _rfc822.decode(string), session
-    except DecodeError:
-        pass
-    try:
-        return _rfc3339.decode(string), session
-    except DecodeError:
-        pass
-    for fmt, tzinfo in _datetime_formats:
-        try:
-            if IRON_PYTHON:
-                # IronPython strptime() seems to ignore whitespace
-                string = string.replace(' ', '|')
-                fmt = fmt.replace(' ', '|')
-            if fmt.endswith('%z'):
-                dt = datetime.datetime.strptime(string[:-5], fmt[:-2])
-                tz_sign = -1 if string[-5:-4] == '-' else 1
-                tz_hour = int(string[-4:-2])
-                tz_min = int(string[-2:])
-                tzinfo = FixedOffset(tz_sign * (tz_hour * 60 + tz_min))
-            else:
-                dt = datetime.datetime.strptime(string, fmt)
-            return dt.replace(tzinfo=tzinfo or session.default_tz_info), session
-        except ValueError:
-            continue
-    raise ValueError('failed to parse datetime: ' + repr(string))
+    return datetime_parser(element, session)
 
 
 @parse_channel.path('managingEditor', attr_name='contributors')
 @parse_channel.path('webMaster', attr_name='contributors')
 @parse_item.path('author', attr_name='authors')
 def parse_person(element, session):
-    string = element.text
-    name, email_addr = email.utils.parseaddr(string)
-    if '@' not in email_addr:
-        if not name:
-            name = email_addr
-        email_addr = None
-    if not name:
-        name = email_addr
-    if not name:
-        return None, session
-    person = Person(name=name, email=email_addr or None)
-    return person, session
+    return person_parser(element, session)
 
 
 @parse_channel.path('category', attr_name='categories')
@@ -136,17 +66,12 @@ def parse_category(element, session):
 @parse_channel.path('copyright', attr_name='rights')
 @parse_item.path('title')
 def parse_text(element, session):
-    return Text(value=element.text or ''), session
+    return text_parser(element, session)
 
 
 @parse_channel.path('description', attr_name='subtitle')
 def parse_subtitle(element, session):
-    return Text(type='text', value=element.text), session
-
-
-@parse_item.path(CONTENT_XMLNS + 'encoded', 'content')
-def parse_content(element, session):
-    return Content(type='html', value=element.text), session
+    return subtitle_parser(element, session)
 
 
 @parse_channel.path('link', ATOM_XMLNS_SET, attr_name='links')
@@ -160,12 +85,7 @@ def parse_atom_link(element, session):
 @parse_channel.path('link', attr_name='links')
 @parse_item.path('link', attr_name='links')
 def parse_link(element, session):
-    if not element.text:
-        return None, session
-    link = Link(uri=element.text,
-                relation='alternate',
-                mimetype='text/html')
-    return link, session
+    return link_parser(element, session)
 
 
 @parse_channel.path('generator')
@@ -208,7 +128,7 @@ def parse_comments(element, session):
 @parse_item.path('description', attr_name='content')
 @parse_item.path('encoded', [CONTENT_XMLNS], 'content')
 def parse_content(element, session):
-    return Content(type='html', value=element.text), session
+    return content_parser(element, session)
 
 
 @parse_item.path('guid', attr_name='id')
@@ -221,24 +141,7 @@ def parse_guid(element, session):
     return None, session
 
 
-def guess_default_tzinfo(root, url):
-    """Guess what time zone is implied in the feed by seeing the TLD of
-    the ``url`` and its ``<language>`` tag.
-
-    """
-    lang = root.find('channel/language')
-    if lang is None or not lang.text:
-        return utc
-    lang = lang.text.strip()
-    if len(lang) == 5 and lang[2] == '-':
-        lang = lang[:2]
-    parsed = urlparse.urlparse(url)
-    domain = parsed.hostname.rsplit('.', 1)
-    country = domain[1] if len(domain) > 1 and len(domain[1]) == 2 else None
-    return guess_tzinfo_by_locale(lang, country) or utc
-
-
-def parse_rss(xml, feed_url=None, parse_entry=True):
+def parse_rss2(xml, feed_url=None, parse_entry=True):
     """Parse RSS 2.0 XML and translate it into Atom.
 
     To make the feed data valid in Atom format, ``id`` and ``link[rel=self]``
@@ -260,7 +163,7 @@ def parse_rss(xml, feed_url=None, parse_entry=True):
     root = fromstring(normalize_xml_encoding(xml))
     channel = root.find('channel')
     default_tzinfo = guess_default_tzinfo(root, feed_url)
-    session = RSS2Session(feed_url, default_tzinfo)
+    session = RSSSession(feed_url, default_tzinfo)
     feed_data = parse_channel(channel, session)
     if parse_entry:
         items = channel.findall('item')
@@ -268,35 +171,5 @@ def parse_rss(xml, feed_url=None, parse_entry=True):
         for item in items:
             entry_list.append(parse_item(item, session))
         feed_data.entries = entry_list
-    check_valid_as_atom(feed_data, session)
+    make_legal_as_atom(feed_data, session)
     return feed_data, None
-
-
-def check_valid_as_atom(feed_data, session):
-    # FIXME: It doesn't only "check" the feed_data but manipulates it
-    # if not valid.  I think the function should be renamed.
-    if not feed_data.id:
-        feed_data.id = session.feed_url
-    if all(l.relation != 'self' for l in feed_data.links):
-        feed_data.links.insert(0, Link(relation='self', uri=session.feed_url))
-    for entry in feed_data.entries:
-        if entry.updated_at is None:
-            entry.updated_at = entry.published_at
-        if entry.id is None:
-            entry.id = entry.links[0].uri if entry.links else ''
-    if feed_data.updated_at is None:
-        if feed_data.entries:
-            try:
-                feed_data.updated_at = max(entry.updated_at
-                                           for entry in feed_data.entries
-                                           if entry.updated_at)
-            except ValueError:
-                feed_data.updated_at = now()
-                for entry in feed_data.entries:
-                    if entry.updated_at is None:
-                        entry.updated_at = feed_data.updated_at
-        else:
-            feed_data.updated_at = now()
-    if feed_data.title is None:
-        feed_data.title = feed_data.subtitle
-        # FIXME: what should we do when there's even no subtitle?